Move Add/Remove/Update options to DrinkService
