--- src/main/java/com/infoshareacademy/service/DatabaseOperator.java
@@ -1,5 +1,0 @@
-package com.infoshareacademy.service;
-
-public interface DatabaseOperator {
-    boolean operate(String drinkId);
-}
